requirements.txt
fastapi>=0.97.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
slowapi>=0.1.7
pillow>=9.5.0
opencv-python>=4.8.0
replicate>=0.11.0
httpx>=0.24.1
python-multipart>=0.0.6
imageio>=2.31.1
imageio-ffmpeg>=0.4.8
numpy>=1.24.3